+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpService.java
@@ -55,6 +55,7 @@ public class HttpService {
     private static final BString CORS_FIELD = StringUtils.fromString("cors");
     private static final BString VERSIONING_FIELD = StringUtils.fromString("versioning");
     private static final BString HOST_FIELD = StringUtils.fromString("host");
+    private static final BString MEDIA_TYPE_SUBTYPE_PREFIX = StringUtils.fromString("mediaTypeSubtypePrefix");
 
     private BObject balService;
     private List<HttpResource> resources;
@@ -67,6 +68,7 @@ public class HttpService {
     private BMap<BString, Object> compression;
     private String hostName;
     private String chunkingConfig;
+    private String mediaTypeSubtypePrefix;
 
     protected HttpService(BObject service, String basePath) {
         this.balService = service;
@@ -138,6 +140,14 @@ public class HttpService {
         return hostName;
     }
 
+    public void setMediaTypeSubtypePrefix(String mediaTypeSubtypePrefix) {
+        this.mediaTypeSubtypePrefix = mediaTypeSubtypePrefix;
+    }
+
+    public String getMediaTypeSubtypePrefix() {
+        return mediaTypeSubtypePrefix;
+    }
+
     public String getBasePath() {
         return basePath;
     }
@@ -185,6 +195,10 @@ public class HttpService {
             httpService.setChunkingConfig(serviceConfig.get(HttpConstants.ANN_CONFIG_ATTR_CHUNKING).toString());
             httpService.setCorsHeaders(CorsHeaders.buildCorsHeaders(serviceConfig.getMapValue(CORS_FIELD)));
             httpService.setHostName(serviceConfig.getStringValue(HOST_FIELD).getValue().trim());
+            if (serviceConfig.containsKey(MEDIA_TYPE_SUBTYPE_PREFIX)) {
+                httpService.setMediaTypeSubtypePrefix(serviceConfig.getStringValue(MEDIA_TYPE_SUBTYPE_PREFIX)
+                        .getValue().trim());
+            }
         } else {
             httpService.setHostName(HttpConstants.DEFAULT_HOST);
         }